Customize the Template for our needs: iOS + Swift
Delete .travis.yml

Update ConfigureiOS.rb

Update template to reflect library creation at ASDA

Fix tests path

Organize the file structure for the Swift template

Update project example template

fix project example path

fix project example scheme name

Update Podfile config and License files

Add project to Podfile template

Update template structure

Fix paths

Fix paths for podfiles

Fix rename command

fix rename command

fix mv command

fix mv command

fix scheme rename

fix paths

fix path

Move files in template

Fix paths and template structure

Add question to decide whether or not to create a git repo

add create git option

fix template for no demo target

check if demo target is selected

Default to iOS and Swift

fix configurator when no demo project

fix project configurator when demo target

Change defautls to be No

remove Pod folder

update podspec template

update license in podspec template

+++ setup/TemplateConfigurator.rb
@@ -70,13 +70,15 @@ module Pod
     def run
       @message_bank.welcome_message
 
-      platform = self.ask_with_answers("What platform do you want to use?", ["iOS", "macOS"]).to_sym
+      # platform = self.ask_with_answers("What platform do you want to use?", ["iOS", "macOS"]).to_sym
+      platform = :ios
 
       case platform
         when :macos
           ConfigureMacOSSwift.perform(configurator: self)
         when :ios
-          framework = self.ask_with_answers("What language do you want to use?", ["Swift", "ObjC"]).to_sym
+          # framework = self.ask_with_answers("What language do you want to use?", ["Swift", "ObjC"]).to_sym
+          framework = :swift
           case framework
             when :swift
               ConfigureSwift.perform(configurator: self)
@@ -89,11 +91,11 @@ module Pod
       replace_variables_in_files
       clean_template_files
       rename_template_files
-      add_pods_to_podfile
+      add_pods_to_podfile(podfile_path)
+      add_pods_to_podfile(example_podfile_path) if File.exists?(example_podfile_path)
       customise_prefix
-      rename_classes_folder
       ensure_carthage_compatibility
-      reinitialize_git_repo
+      # reinitialize_git_repo
       run_pod_install
 
       @message_bank.farewell_message
@@ -102,19 +104,25 @@ module Pod
     #----------------------------------------#
 
     def ensure_carthage_compatibility
-      FileUtils.ln_s('Example/Pods/Pods.xcodeproj', '_Pods.xcodeproj')
+      FileUtils.ln_s('Pods/Pods.xcodeproj', '_Pods.xcodeproj')
     end
 
     def run_pod_install
       puts "\nRunning " + "pod install".magenta + " on your new library."
       puts ""
 
-      Dir.chdir("Example") do
-        system "pod install"
+      system "pod install"
+
+      if File.exists?("Example")
+        puts "\nRunning " + "pod install".magenta + " on your new Example project."
+        puts ""
+        Dir.chdir("Example") do
+          system "pod install"
+        end
       end
 
-      `git add Example/#{pod_name}.xcodeproj/project.pbxproj`
-      `git commit -m "Initial commit"`
+      # `git add Example/#{pod_name}-Example.xcodeproj/project.pbxproj`
+      # `git commit -m "Initial commit"`
     end
 
     def clean_template_files
@@ -124,16 +132,18 @@ module Pod
     end
 
     def replace_variables_in_files
-      file_names = ['POD_LICENSE', 'POD_README.md', 'NAME.podspec', '.travis.yml', podfile_path]
+      file_names = ['POD_LICENSE', 'POD_README.md', 'NAME.podspec', podfile_path, example_podfile_path]
       file_names.each do |file_name|
-        text = File.read(file_name)
-        text.gsub!("${POD_NAME}", @pod_name)
-        text.gsub!("${REPO_NAME}", @pod_name.gsub('+', '-'))
-        text.gsub!("${USER_NAME}", user_name)
-        text.gsub!("${USER_EMAIL}", user_email)
-        text.gsub!("${YEAR}", year)
-        text.gsub!("${DATE}", date)
-        File.open(file_name, "w") { |file| file.puts text }
+        if File.exists?(file_name)
+          text = File.read(file_name)
+          text.gsub!("${POD_NAME}", @pod_name)
+          text.gsub!("${REPO_NAME}", @pod_name.gsub('+', '-'))
+          text.gsub!("${USER_NAME}", user_name)
+          text.gsub!("${USER_EMAIL}", user_email)
+          text.gsub!("${YEAR}", year)
+          text.gsub!("${DATE}", date)
+          File.open(file_name, "w") { |file| file.puts text }
+        end
       end
     end
 
@@ -141,13 +151,13 @@ module Pod
       @pods_for_podfile << podname
     end
 
-    def add_pods_to_podfile
-      podfile = File.read podfile_path
+    def add_pods_to_podfile(file_path)
+      podfile = File.read file_path
       podfile_content = @pods_for_podfile.map do |pod|
         "pod '" + pod + "'"
       end.join("\n    ")
       podfile.gsub!("${INCLUDED_PODS}", podfile_content)
-      File.open(podfile_path, "w") { |file| file.puts podfile }
+      File.open(file_path, "w") { |file| file.puts podfile }
     end
 
     def add_line_to_pch line
@@ -155,7 +165,7 @@ module Pod
     end
 
     def customise_prefix
-      prefix_path = "Example/Tests/Tests-Prefix.pch"
+      prefix_path = "Tests/Tests-Prefix.pch"
       return unless File.exists? prefix_path
 
       pch = File.read prefix_path
@@ -165,7 +175,7 @@ module Pod
 
     def set_test_framework(test_type, extension, folder)
       content_path = "setup/test_examples/" + test_type + "." + extension
-      tests_path = "templates/" + folder + "/Example/Tests/Tests." + extension
+      tests_path = "templates/" + folder + "/Tests/Tests." + extension
       tests = File.read tests_path
       tests.gsub!("${TEST_EXAMPLE}", File.read(content_path) )
       File.open(tests_path, "w") { |file| file.puts tests }
@@ -175,10 +185,6 @@ module Pod
       FileUtils.mv "POD_README.md", "README.md"
       FileUtils.mv "POD_LICENSE", "LICENSE"
       FileUtils.mv "NAME.podspec", "#{pod_name}.podspec"
-    end
-
-    def rename_classes_folder
-      FileUtils.mv "Pod", @pod_name
     end
 
     def reinitialize_git_repo
@@ -216,6 +222,10 @@ module Pod
     end
 
     def podfile_path
+      'Podfile'
+    end
+
+    def example_podfile_path
       'Example/Podfile'
     end
